exports progress bar on both recipe and collection forms
