+++ src/components/forms/UploadCollections/UploadCollection_description.js
@@ -1,17 +1,11 @@
 import React from 'react';
 import { useHistory } from 'react-router-dom';
-import { makeStyles, Container, Typography, Button, TextField, Grid, Box } from '@material-ui/core';
+import { makeStyles, Container, Typography, Button, TextField, Grid } from '@material-ui/core';
 import Icon from '../Icon';
-import BorderLinearProgress from '../BorderLinearProgress'
+import ProgressBar from '../ProgressBar';
 import CancelButton from './CancelButton';
 
 const useStyles = makeStyles((theme) => ({
-    bar: {
-        display: "flex", 
-        alignItems: "center",
-        flexGrow: 1,
-        margin: theme.spacing(3, 0)
-    },
     paper: {
         display: "flex",
         flexDirection: "column",
@@ -99,11 +93,7 @@ const UploadCollection_description = ({ collectionData, setCollectionData, handl
                     </Button>
                 </Grid>
             </Grid>
-            <Box className={classes.bar}>
-                <Box className={classes.fullWidth} mr={1}>
-                    <BorderLinearProgress variant="determinate" value={66} />
-                </Box>
-            </Box>
+            <ProgressBar num={66}/>
            <CancelButton handleCancel={handleCancel}/>
         </Container>
     );
