+++ src/components/forms/UploadRecipes/UploadRecipe_name.js
@@ -1,16 +1,11 @@
 import React from 'react';
 import { useHistory } from 'react-router-dom';
-import { makeStyles, Container, Typography, Button, TextField, Grid, Box } from '@material-ui/core';
-import BorderLinearProgress from '../BorderLinearProgress';
+import { makeStyles, Container, Typography, Button, TextField, Grid } from '@material-ui/core';
+import ProgressBar from '../ProgressBar';
 import CancelButton from './CancelButton';
 import Icon from '../Icon';
 
 const useStyles = makeStyles((theme) => ({
-    bar: {
-        display: "flex", 
-        flexGrow: 1,
-        margin: theme.spacing(3, 0)
-    },
     paper: {
         display: "flex",
         flexDirection: "column",
@@ -97,11 +92,7 @@ const UploadRecipe = ({ handleCancel, recipeData, setRecipeData }, props) => {
                     </Button>
                 </Grid>
             </Grid>  
-            <Box className={classes.bar}>
-                <Box className={classes.fullWidth} mr={1}>
-                    <BorderLinearProgress variant="determinate" value={9} />
-                </Box>
-            </Box>
+            <ProgressBar num={9}/>
             <CancelButton handleCancel={handleCancel}/>
         </Container>
     );
