+++ src/components/forms/UploadRecipes/UploadRecipe_pairings.js
@@ -1,16 +1,11 @@
 import React from 'react';
 import { useHistory } from 'react-router-dom';
-import { makeStyles, Container, Typography, Button, TextField, Grid, Box } from '@material-ui/core';
+import { makeStyles, Container, Typography, Button, TextField, Grid } from '@material-ui/core';
 import Icon from '../Icon';
-import BorderLinearProgress from '../BorderLinearProgress';
+import ProgressBar from '../ProgressBar';
 import CancelButton from './CancelButton';
 
 const useStyles = makeStyles((theme) => ({
-    bar: {
-        display: "flex", 
-        flexGrow: 1,
-        margin: theme.spacing(3, 0)
-    },
     paper: {
         display: "flex",
         flexDirection: "column",
@@ -115,11 +110,7 @@ const UploadRecipe_pairings = ({ handleCancel, setRecipeData, recipeData }) => {
                     </Button>
                 </Grid>
             </Grid>
-            <Box className={classes.bar}>
-                <Box className={classes.fullWidth} mr={1}>
-                    <BorderLinearProgress variant="determinate" value={90} />
-                </Box>
-            </Box>
+            <ProgressBar num={90}/>
             <CancelButton handleCancel={handleCancel}/>
         </Container>
     );
